+++ lib/features/auth/presentation/pages/delete_account/delete_account_screen.dart
@@ -6,7 +6,7 @@ import 'package:goms/core/theme/colors/app_colors.dart';
 import 'package:goms/features/auth/presentation/pages/auth_base_screen.dart';
 import 'package:goms/features/auth/presentation/pages/delete_account/models/delete_account_state.dart';
 import 'package:goms/features/auth/presentation/pages/delete_account/viewModels/delete_account_provider.dart';
-import 'package:goms/core/widgets/common/goms_dialog.dart';
+import 'package:goms/core/widgets/common/dialogs/goms_dialog.dart';
 import 'package:goms/core/widgets/common/text_fields/password_text_field.dart';
 
 class DeleteAccountScreen extends ConsumerStatefulWidget {
