+++ lib/features/auth/presentation/pages/reset_password/reset_password_screen.dart
@@ -9,7 +9,7 @@ import 'package:goms/core/theme/typography/app_text_styles.dart';
 import 'package:goms/features/auth/presentation/pages/auth_base_screen.dart';
 import 'package:goms/features/auth/presentation/pages/reset_password/models/reset_password_state.dart';
 import 'package:goms/features/auth/presentation/pages/reset_password/viewModels/reset_password_provider.dart';
-import 'package:goms/core/widgets/common/goms_dialog.dart';
+import 'package:goms/core/widgets/common/dialogs/goms_dialog.dart';
 import 'package:goms/core/widgets/common/text_fields/password_text_field.dart';
 
 class ResetPasswordScreen extends ConsumerStatefulWidget {
